Revert "Change Custody Count to Uint8 (#14386)" (#14415)
This reverts commit bd7ec3fa976c49cbe67c16f9f20cb8cfcaf882b2.

+++ proto/prysm/v1alpha1/metadata/metadata_interfaces.go
@@ -11,7 +11,7 @@ type Metadata interface {
 	SequenceNumber() uint64
 	AttnetsBitfield() bitfield.Bitvector64
 	SyncnetsBitfield() bitfield.Bitvector4
-	CustodySubnetCount() uint8
+	CustodySubnetCount() uint64
 	InnerObject() interface{}
 	IsNil() bool
 	Copy() Metadata
